deleted old unnecessary files and created legacy folder for old Island class
