--- bingo/EA/test_var_or.py
@@ -1,34 +1,0 @@
-import pytest
-import numpy as np
-
-from bingo.Base.Variation import Variation
-from bingo.MultipleValues import *
-from VarOr import *
-
-
-@pytest.fixture
-def population():
-    generator = MultipleValueGenerator(mutation_function, 10)
-    return [generator() for i in range(25)]
-
-@pytest.fixture
-def var_or():
-    crossover = SinglePointCrossover()
-    mutation = SinglePointMutation(mutation_function)
-    var_Or = VarOr(crossover, mutation, 0.2, 0.4)
-    return var_Or
-
-def mutation_function():
-    return np.random.choice([True, False])
-
-def test_offspring_not_equals_parents(population, var_or):
-    offspring = var_or(population, 25)
-    assert offspring != population 
-    assert offspring is not population
-        
-
-def test_no_two_variations_at_once(population, var_or):
-    offspring = var_or(population, 25)
-    for i, indv in enumerate(var_or.crossover_offspring):
-        assert not (indv and var_or.mutation_offspring[i])
-
